+++ CHANGELOG.md
@@ -5,11 +5,18 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Added `FanOutRequestCollapser.Builder#withGroupId`, so two instance can be differentiated when observing the logs.
+- Added `FanOutRequestCollapser.Builder#withBatchMaxConcurrencyWaitTime` to set the max wait time for a full bulkhead to free up before making the next call failed.
+### Changed
+- Made `FanOutRequestCollapser#maxConcurrency` limit forced by `Bulkhead` instead of the concurrency of `flatMap`,
+  which killed the whole collapser instead of that single call over the limit.
+- Dropped support of reactor-core below 3.4.0, by using the new `Sinks` api.
 
 ## [1.1.3]
 ### Changed
 - Migrated most of the unit tests to JUnit 5.
-- Extended testing of ReactiveCache implementations.
+- Extended testing of `ReactiveCache` implementations.
 
 ## [1.1.2]
 ### Added