+++ CHANGELOG.md
@@ -12,6 +12,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Made `FanOutRequestCollapser#maxConcurrency` limit forced by `Bulkhead` instead of the concurrency of `flatMap`,
   which killed the whole collapser instead of that single call over the limit.
 - Dropped support of reactor-core below 3.4.0, by using the new `Sinks` api.
+### Fixed
+- Fixed `ReactiveCache` implementations to log the `Throwable#toString()` instead of the message, which can be null.
 
 ## [1.1.3]
 ### Changed